+++ brage-import/src/main/java/no/sikt/nva/brage/migration/merger/CristinImportPublicationMerger.java
@@ -6,6 +6,12 @@ import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
+import no.sikt.nva.brage.migration.merger.publicationcontextmerger.AnthologyMerger;
+import no.sikt.nva.brage.migration.merger.publicationcontextmerger.BookMerger;
+import no.sikt.nva.brage.migration.merger.publicationcontextmerger.DegreeMerger;
+import no.sikt.nva.brage.migration.merger.publicationcontextmerger.EventMerger;
+import no.sikt.nva.brage.migration.merger.publicationcontextmerger.JournalMerger;
+import no.sikt.nva.brage.migration.merger.publicationcontextmerger.ReportMerger;
 import no.sikt.nva.brage.migration.model.PublicationRepresentation;
 import no.unit.nva.model.AdditionalIdentifier;
 import no.unit.nva.model.Contributor;
