+++ brage-import/src/main/java/no/sikt/nva/brage/migration/merger/publicationcontextmerger/JournalMerger.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.brage.migration.merger;
+package no.sikt.nva.brage.migration.merger.publicationcontextmerger;
 
 import static java.util.Objects.nonNull;
 import no.unit.nva.model.contexttypes.Journal;
