+++ brage-import/src/main/java/no/sikt/nva/brage/migration/merger/publicationcontextmerger/PublicationContextMerger.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.brage.migration.merger;
+package no.sikt.nva.brage.migration.merger.publicationcontextmerger;
 
 import static java.util.Objects.nonNull;
 import java.util.List;
